+++ packages/acme_chart/lib/src/add_ons/drawing_tools_ui/continuous/continuous_drawing_tool_config.dart
@@ -37,6 +37,7 @@ class ContinuousDrawingToolConfig extends DrawingToolConfig {
         ..putIfAbsent(DrawingToolConfig.nameKey, () => name);
 
   /// Drawing tool line style
+  @override
   final LineStyle lineStyle;
 
   /// Drawing tool line pattern: 'solid', 'dotted', 'dashed'