+++ packages/acme_chart/lib/src/add_ons/drawing_tools_ui/doodle/doodle_drawing_tool_config.dart
@@ -41,6 +41,7 @@ class DoodleDrawingToolConfig extends DrawingToolConfig {
         ..putIfAbsent(DrawingToolConfig.nameKey, () => name);
 
   /// Drawing tool line style
+  @override
   final LineStyle lineStyle;
 
   @override