+++ packages/acme_chart/lib/src/add_ons/drawing_tools_ui/line/line_drawing_tool_config_mobile.dart
@@ -39,6 +39,7 @@ class LineDrawingToolConfigMobile extends DrawingToolConfig {
         ..putIfAbsent(DrawingToolConfig.nameKey, () => name);
 
   /// Drawing tool line style
+  @override
   final LineStyle lineStyle;
 
   /// Drawing tool overlay style