+++ packages/acme_chart/lib/src/add_ons/drawing_tools_ui/segment/segment_drawing_tool_config.dart
@@ -42,6 +42,7 @@ class SegmentDrawingToolConfig extends DrawingToolConfig {
         ..putIfAbsent(DrawingToolConfig.nameKey, () => name);
 
   /// Drawing tool line style
+  @override
   final LineStyle lineStyle;
 
   /// Drawing tool line pattern: 'solid', 'dotted', 'dashed'