+++ packages/acme_chart/lib/src/add_ons/drawing_tools_ui/trend/trend_drawing_tool_config.dart
@@ -86,6 +86,7 @@ class TrendDrawingToolConfig extends DrawingToolConfig {
   final LineStyle fillStyle;
 
   /// Drawing tool line style
+  @override
   final LineStyle lineStyle;
 
   /// Drawing tool line pattern: 'solid', 'dotted', 'dashed'